pyproject.toml
[tool.poetry]
name = "home-credit-default-risk"
version = "0.1.0"
description = ""
authors = ["kou1218 <koudai141218@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
pandas = "^2.1.3"
xgboost = "^2.0.2"
hydra-core = "^1.3.2"
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
